refactor: merge files and simplify code
